Remove git dependency.
The plugins run git describe in a subprocess to determine the
plugin version, which is very error prone inside the snap. This commit
runs the same git describe command from the snap build and then patches
the source code to return this value, so that the plugins no longer need
to run git commands from inside the snap.